Improve infrastructure - Better rollup+ts config - Remove unnecesary type generation (faster builds) - Rename tests
