+++ packages/chord-detect/package.json
@@ -8,12 +8,12 @@
     "theory",
     "@tonaljs/tonal"
   ],
-  "main": "dist/index.es5.js",
-  "module": "dist/index.esnext.js",
+  "main": "dist/index.js",
+  "module": "dist/index.es.js",
   "files": [
     "dist"
   ],
-  "types": "dist/chord-detect/index.d.ts",
+  "types": "dist/index.dt.ts",
   "dependencies": {
     "@tonaljs/chord-type": "^3.5.1",
     "@tonaljs/core": "^3.5.0",
