+++ packages/chord-dictionary/package.json
@@ -3,12 +3,12 @@
   "version": "3.4.3",
   "description": "A dictionary of musical chords",
   "keywords": [],
-  "main": "dist/index.es5.js",
-  "module": "dist/index.esnext.js",
+  "main": "dist/index.js",
+  "module": "dist/index.es.js",
   "files": [
     "dist"
   ],
-  "types": "dist/chord-dictionary/index.d.ts",
+  "types": "dist/index.dt.ts",
   "dependencies": {
     "@tonaljs/chord-type": "^3.5.1"
   },
