+++ packages/chord-type/package.json
@@ -10,12 +10,12 @@
     "theory",
     "@tonaljs/core"
   ],
-  "main": "dist/index.es5.js",
-  "module": "dist/index.esnext.js",
+  "main": "dist/index.js",
+  "module": "dist/index.es.js",
   "files": [
     "dist"
   ],
-  "types": "dist/chord-type/index.d.ts",
+  "types": "dist/index.dt.ts",
   "dependencies": {
     "@tonaljs/core": "^3.5.0",
     "@tonaljs/pcset": "^3.5.0"
