+++ packages/core/package.json
@@ -9,12 +9,12 @@
     "music",
     "theory"
   ],
-  "main": "dist/index.es5.js",
-  "module": "dist/index.esnext.js",
+  "main": "dist/index.js",
+  "module": "dist/index.es.js",
   "files": [
     "dist"
   ],
-  "types": "dist/core/index.d.ts",
+  "types": "dist/index.dt.ts",
   "author": "danigb@gmail.com",
   "license": "MIT",
   "publishConfig": {
