+++ packages/interval/test.ts
