+++ packages/modules/package.json
@@ -3,12 +3,12 @@
   "version": "3.4.10",
   "description": "deprecated",
   "keywords": [],
-  "main": "dist/index.es5.js",
-  "module": "dist/index.esnext.js",
+  "main": "dist/index.js",
+  "module": "dist/index.es.js",
   "files": [
     "dist"
   ],
-  "types": "dist/modules/index.d.ts",
+  "types": "dist/index.dt.ts",
   "dependencies": {
     "@tonaljs/tonal": "^3.6.0"
   },
