+++ packages/range/package.json
@@ -8,12 +8,12 @@
     "theory",
     "@tonaljs/core"
   ],
-  "main": "dist/index.es5.js",
-  "module": "dist/index.esnext.js",
+  "main": "dist/index.js",
+  "module": "dist/index.es.js",
   "files": [
     "dist"
   ],
-  "types": "dist/range/index.d.ts",
+  "types": "dist/index.dt.ts",
   "dependencies": {
     "@tonaljs/collection": "^3.5.0",
     "@tonaljs/midi": "^3.5.0"
