+++ packages/tonal/rollup.browser.config.js
@@ -4,21 +4,21 @@ import { terser } from "rollup-plugin-terser";
 import filesize from "rollup-plugin-filesize";
 
 export default {
-  input: "./dist/index.es5.js",
+  input: "./dist/index.js",
   output: {
     format: "iife",
     file: "./browser/tonal.min.js",
     name: "Tonal",
-    sourcemap: true,
+    sourcemap: true
   },
   plugins: [
     resolve({
       jsnext: true,
       main: true,
-      browser: true,
+      browser: true
     }),
     commonjs(),
     terser(),
-    filesize(),
-  ],
+    filesize()
+  ]
 };
